+++ src/main/java/com/boleg/crud_pro/dao/UserDAO.java
@@ -10,9 +10,9 @@ public interface UserDAO {
 
     public void saveUser(User user);
 
-    public User getUserById(int id);
+    public User getUserById(Long id);
 
-    public void deleteUserById(int id);
+    public void deleteUserById(Long id);
 
     public User getUserByEmail(String email);
 }
